+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.0] - 2025-07-28
 
+### Added
+
+- Add update win/loss/total match flow after game completion
+
 ### Fixed
 
 - Correct scoring flow during round submission